wayfound-backend/requirements.txt
# Core FastAPI and GraphQL
fastapi==0.104.1
strawberry-graphql[fastapi]==0.213.0
uvicorn[standard]==0.24.0

# Database - Fixed compatibility issues
asyncpg==0.29.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM and AI - Fixed compatibility
openai==1.45.0
langchain==0.2.16
langchain-openai==0.1.23
langchain-core==0.2.38

# HTTP and async
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1